+++ NEWS.md
@@ -1,7 +1,22 @@
+# mikropml 1.0.0
+
+- mikropml now has a logo created by @NLesniak!
+- Made documentation improvements (#238, #231 @kelly-sovacool; #256 @BTopcuoglu).
+- New option in `preprocess_data()`: `prefilter_threshold` (#240, @kelly-sovacool, @courtneyarmour).
+    - Remove any features that appear in N=`prefilter_threshold` or fewer rows in the data.
+    - Created function `remove_singleton_columns()` called by `preprocess_data()` to carry this out.
+- New option in `get_feature_importance()`: `groups` (#246, @kelly-sovacool).
+    - Provide custom groups of features to permute together during permutation importance.
+    - `groups` is `NULL` by default; in this case, correlated features above `corr_thresh` are grouped together.
+- `preprocess_data()` now replaces spaces in the outcome column with underscores (#247, @kelly-sovacool, @JonnyTran).
+- Clarify in the intro vignette that we do not support multi-label outcomes. (#254, @zenalapp)
+- Optional progress bar for `preprocess_data()` and `get_feature_importance()` using [the progressr package](https://github.com/HenrikBengtsson/progressr) (#257, @kelly-sovacool, @JonnyTran, @FedericoComoglio).
+- The mikropml paper is soon to be published in [JOSS](https://joss.theoj.org/papers/72bf31a3f51f8fc273ef6b99bd04ede1)!
+
 # mikropml 0.0.2
 
 - Fixed a test failure on Solaris.
-- Fixed multiple test failtures with R 3.6.2 due to `stringsAsFactors` behavior.
+- Fixed multiple test failures with R 3.6.2 due to `stringsAsFactors` behavior.
 - Made minor documentation improvements.
 - Moved `rpart` from Suggests to Imports for consistency with other packages used during model training.